Phase 4d: extract useCalendarDataPipeline hook
Moves the entire data pipeline (~190 lines) into src/hooks/useCalendarDataPipeline.ts:
fetch, source aggregation, Supabase realtime, event normalization, CalendarEngine,
engine sync effects, base/region/label derivation, VIEWS computation, scoped events,
categories, filter/sort pipeline, and on-shift IDs.

WorksCalendar.tsx calls the hook in ~22 lines and removes 14 unused imports.

1270 → 1088 lines.
